Fixes to get a lot more postgres tests cleared. The primary issue that was causing a bunch of advanced tests to fail was the default ordering of jobs returned (when no ordering was specified). SQLite and Postgres return different default ordering. The solution to make the tests robust was the enforce a default order (when none is specified) of Job.start. This reverses from our earlier defaults, and I've updated the docs.
More tests remain to be fixed for postgres, but this sets up a
solid foundation. I spent a bunch of time dealing with tags ordering
in op_metrics, only to track the issue down to the changed default
ordering. Now the tests in test_query will run in the ordering where
simpler tests will fail first.