- for EPMT Query and Outliers API added safeguards to prevent avoidable   querying across the database. We require functions to explicitly   specify jobs for most functions and will not accept an empty   list. For functions such as get_jobs and get_procs we apply sensible   defaults for limit.
The above applies to all formats other than ORM. ORM is assumed
to be an advanced usage, and it's assumed the user knows what they're
doing. Further, orm queries are lazily-evaluated and can be chained,
so it's not desirable to constrain them unnecessarily. It's quite
possible chaining subsequent filters would limit the results anyway.