Added epmt-dash as a submodule. This will allow dash to easily use epmt_query.py and epmt_outliers.py.
To get epmt-dash populated in an existing EPMT source hierarchy:

$ git checkout sow3
$ git merge origin/sow3

At this point you will have a `.gitmodules` in your root directory.
You can inspect it if you like. It's a text file.

To populate the submodule directories:

$ git submodule init
$ git submodule update

$ cd dash
$ ls

You will notice that .gitmodules is set to track the multi_page branch
of epmt-dash.