Numerous minor fixes to get tests to pass with changes in the ingestion pipeline:
  The fixes were mostly trivial in that we have removed the synthetic
metric 'user+system', and so we have one less dataframe columns. This
accounts for a majority of the unit test fixes.

  The other fix was a more important change to how the threads dataframe
is saved. Earlier it used to be saved as a split json. This was time-consuming
to derive given than we have a list of thread dicts. So instead we now directly
save the list of thread dicts. That is faster for ingestion as well. As a bonus,
the conversion of list of thread dicts to a dataframe is a standard and
faster method pd.DataFrame(list_of_thread_dicts). It's important to note the
API interface hasn't changed and our API calls still get the same threads dataframe
on querying.