- High-level integration of MVOD into our detect_outlier_jobs.   This code is untested, and will need some cleanup and fixes   before it's all done.
There is a small chance that this commit breaks the CI pipeline
since I did make significant changes to detect_outlier_jobs.

MVOD for ops is still ways off. One reason we can't just drop-in the
MV classifiers is the way we iterate over features. The univariate
code loops over inidividual features calling the detector repeatedly
for each feature vector. The MVOD has to be called once for the
entire feature set. There are also some other subtle differences
over how and what needs to be stored and retrieved from the model,
which is different between univariate detectors and mv detectors.