addressed an important complication when using PCA-based trained models:     Prior to doing PCA, we need also need to include the model jobs     along with the passed jobs. This is necessary as PCA requires     scaling to be performed and so needs all the jobs.
We now check if any model jobs are not included in the jobs
passed for outlier detection, and if so append them to the
jobs df prior to PCA. Once PCA is done, we remove the just-appended
model jobs, so the outlier detection doesn't show jobs the user
never passed for outlier detection.

All the above only applies to outlier detection using PCA-based
trained models.

At this point we are able to do outlier detection using PCA-based
trained models. Unit tests will be added shortly..