Everything below applies only to SQLAlchemy:
Significant improvements to the SQLAlchemy `setup_db` to ensure
that the schema version is checked and migrations are automatically
applied if needed.

Please note we now have generated a baseline migration file based on
the current state of `models.py`. From now on, we must *never* edit
models.py directly, and instead use `alembic revision` to create
a new migration file for the changes we want. Those schema changes
will be automatically applied when epmt is run (or if you
want, you can just run "alembic upgrade head").

The main accomplishment of this commit is that we now track schema
changes and automatically perform DB schema upgrades if needed.