implemented KMeans clustering with the Silhouette method to identify the optimal number of clusters. This effectively identifies the number of modes in a multimodal distribution.
I want to also add support to identify unimodal distributions using
the Elbow method. We alrady compute the scores, but need to add some
mathematical plumbing to identify the "knee point".

At this point we are properly able to identify the number of modes
in a multimodal distribution (number of modes > 1). I've tested with
a bimodal normal distribution and trimodal.

Need to add unit test case and support to identify unimodal distributions.