- papiex CSV will not contain the jobid as a field. So we save the   staging process table row range in the job info_dict for unprocessed jobs. - schema changes to the staging processes table:   - renamed gen and host_id to match papiex headers (generation and hostname)   - added column in staging table for exitsignal   - made jobid field optional in staging table   - removed threads_sums field from staging table   - removed cpu_time field from staging table
At this point we are getting cat of 100k CSV files into a single
collated file in 1.3s. And we are getting the ingestion of 100k
processes into the staging table in ~ 1.2s

So, all in all we are seeing a performance of around 40k procs/sec
end to end.