Main work in this commit is to support v2 TSV files for import into backends where we don't have direct CSV copy. For this I made very minor changes in get_proc_row() generator function. This way on backends such as SQLite the data is read in and the data structures are created similar to that with CSV v1 format.
The other change is a fix for a bug that surfaced with SQLite when
using ARRAY in a migration file. This was in the processes_staging
table. Since, we don't even use the table in SQLite, I just use the
TEXT format for SQLite (while retaining the ARRAY format for postgres).
Using TEXT is also the suggested workaround for ARRAY in SQLite.