- do not remove metric names from job info_dict, and also   ensure all code-paths (csv v1, v2) have the metric names   saved in the info_dict.
- compile ARRAY as JSON for sqlite.

Unfortunately, using TypeDecorator as suggested in:
http://davidemoro.blogspot.com/2014/10/sqlite-array-type-and-python-sqlalchemy.html

will not work as these don't take arguments. So, these can be used in
a situation where we used ARRAY as ARRAY(). However, postgresql's ARRAY
requires usage as ARRAY(Float).

So, instead we follow the approach suggested in:

https://docs.sqlalchemy.org/en/13/core/custom_types.html

And, we do a compile on SQLite to JSON.