Small change with big ramifications:  - enabled COLLATED_TSV for papiex_options in epmt_default_settings.py
This change will make papiex create collated TSV files. And these
when ingested into EPMT will populate the staging table. Only when
a job is actually queried, and the process data needed, will the job
be processed on-the-fly.

This is a change with big potential for side-effects. So far,
all the tests pass, with one warning in one test:

test_run_nowrap (test.test_run.EPMTShell) ... libpapiex: error open(/tmp/epmt/tushar/1011/perftools.org-pa
piex.tsv): No such file or directory

The issue in this test seems to be the job directory is not created as epmt_start is
predictably not called. Without COLLATED_TSV, we don't get the above warning,
and papiex creates a .csv file. This suggests papiex also creates the
job directory (/tmp/epmt/tushar/1011/) in the earlier code path.

With COLLATED_TSV, we do get a warning from libpapiex; there is no job directory
at the end of the test. And, it seems papiex doesn't create the
job directory (when it doesn't find the job directory). I think this would
suggest a fix needed in papiex for the COLLATED_TSV code path, wherein it
should create the output directory if it doesn't exist.

As epmt_run isn't part of our expected usage I'm going to go ahead
and commit the change to enable COLLATED_TSV by default. Phil,
I'll wait for your comments or call. I'm not making any changes in
papiex so close to the release.