Cleanup stray daemons if the user interrupts "epmt integration".
This problem was slightly tricky because the daemon process is
a grandchild of the parent python process, and has a different process
group. subprocess.run doesn't give the pids, as it waits for the
spawned integration test bats script to finish.

While our daemon does signal handling and cleans up if it gets a
non-fatal signal, I observed that Ctrl-C would only give the top-level
python process the signal. I'm slightly unsure about the last statement,
as processes in the same process group should have got the SIGHUP. Anyhow,
be that as it may, even getting the signal in the bash script did not
result in the daemon getting the signal.

My solution (by no means ideal) was to install handler in the python
script to ennumerate child process pids in the signal handler, and send them
each SIGTERM. This resulted in the bash/bats script receiving a signal.
Therein I installed a trap (in the bash script), to then call "epmt daemon --stop".
This command looks for the PID file in the right place, and sends a
SIGTERM to the daemon to do a clean exit. Two minor things to observe:
Sometimes, the bash script would get the signal too fast, and the daemon
lockfile was not in place. To get around this, I do a loop in the trap
handler to try a few times with a 1sec wait in between.

Also, the daemon had a delay loop of 30sec. This would reduce our throughput,
and also can result in the daemon taking somewhat longer to exit cleanly. I
reduced the daemon delay loop to 10sec. Phil, if you think that's too small,
I'll restore it..

I'd be happier with a cleaner implementation. In my tests, now the cleanup
happens always. But if there's a better way to implement this, I'd gladly
yield!