The main theme in this commit is to skip unit and integration tests that use the daemon for post-processing in cases where the database has unprocessed jobs. We now, check if there are any unprocessed jobs in the database, and if so, we skip these daemon tests.
Along the way, I discovered a bug in epmt_list_jobs, wherein it would return an error code ("epmt list" return a non-zero code), if it got no jobs in the listing. I think this is incorrect, because the database may legitimately have no jobs. For example,

$ epmt list processed=0  # to return unprocessed jobs

should not return an error if there are no unprocessed jobs. It should return an empty list, and a zero eror code. I've fixed this behavior. Feel free to revert this part if you feel otherwise is better..