Fixed a race condition in post-processing, wherein once the job sums were computed and the host-job associations were created using a bulk insert, the execution was interrupted (prior to the orm_commit at the end of the function). The race window is extremely small. To reproduce the bug, you need to add a sleep just before the orm_commit at the end of `post_process_job`, and then do a Ctrl-\ (SIGQUIT).
The reason behind the race condition is that we were bypassing the ORM transactional support by doing the bulk insert of host-job associations. It turns out the bulk insert for the small number of host job associations was anyhow more expensive than using the ORM to do the associations. So, now we do the latter. This has made the code less modal, slightly faster, and now there is no race condtion.