Basic support added to build python3.7.4 using pyenv and then creating a venv prior to runnign make dist. This needs some more work as the python installed using pyenv is missing some shared libraries. I have to figure out how to pass a custom python configure/build argument via pyenv. If that doesn't work, then I'll just use a tarball.
Broadly, the approach I'm taking is that if python3.7.4 is not installed already,
then we will install it using pyenv. Similarly, if the python packages needed
by EPMT are not installed already, then we will create a venv and install them
prior to running pyinstaller. This way, if we want to use the Dockerfile to
build python3 and install the packages, then we will just use that, and not
bother creating a virtual environment. Why use the Dockerfile in preference to
to pyenv? Because, our environment created by Dockerfile is better-controlled,
and that approach is guaranteed to produce,
the right python with all the right shared libraries. Creating python using
pyenv or even via a tarball needs the right packages installed such as
libz-dev, libssl-dev, libsqlite-dev and libreadline-dev. If any of the headers
are not found, then the build skips those pieces, which may lead to a version
of python that might not work fully later. Anyway, I'm working on a separate
branch so we can choose to include this or not based on where we end up.