this fixes an edge case bug with `epmt retire` when the # of jobs to retire is greater than the maximum
the bug occurs when the # of jobs assoc. w/ ref models younger than the
retirement age are spared in a given chunk-deletion. When this occurs,
the retire routine kept re-targeting those assoc jobs for each chunk
deletion after the fact, and kept sparing them.

i.e., if the deletion chunk size were 1000 jobs, and e.g. in the current
chunk, there were 10 jobs to be spared via ref-model association. then
the code would delete 990, leaving the spared-jobs to be the oldest in
the DB, and then the code would try to 1000 of the oldest jobs again,
including the spared jobs.

the solution involves passing an `offset` argument through `retire_jobs`
and to `delete_jobs`. when there are spared jobs, `offset` is incremented
by the difference between the # of deleted jobs, and the # of jobs in the chunk.