Revert "Revert "Merge branch 'feature/7.pip-install' into 'feature/7.pip-install'""
This reverts commit d10fce5d23ea303bcb721642ab78ae39aac667e0.

Apologies to the git history. Awhile back I was perusing my old MR for the retire
bugfix, wanting to look up something in the notes there, and I managed to merge the changes
accidentally. The MR in question is here:
ttps://gitlab.com/minimal-metrics-llc/epmt/epmt/-/merge_requests/13

This wasn't really a big deal, the code was essentially done, I thought I just
shouldn't had done so in the moment, and panic-reverted the merge... that was the real
mistake. As I've learned, reverting merge commits, and then re-doing them, tends to
create a complicated/hard to parse dev history. See:
https://mirrors.edge.kernel.org/pub/software/scm/git/docs/howto/revert-a-faulty-merge.txt

Apparently the best thing to do is to revert the revert commit to re-do
the merge. So that's what is happening now! All I've done here is a
'git revert <commit-hash>', nothing else. For the changes this introduces,
see the MR.