still chasing some tests. forget the dev container for now.
figured out a way to cut down on some of the stranger steps in the build,
see some new env vars in the Makefile at the top. important ones are
output before major steps in the build to aid the builder in what's
going on and what's important.

steps like the installation of bats in /usr/local are now in the
test-release dockerfile. the PATH var is set there now too looking
at /usr/local/bin:/usr/local/libexec for bats explicitly.

something still going on weird with the ops, test_convert_csv still
fails, with the failure seeming to be on convert_csv_in_tar perhaps...